--- src/main/java/Command.java
@@ -1,4 +1,0 @@
-public abstract class Command {
-    abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
-    abstract boolean isExit();
-}
