+++ src/main/java/duke/Parser.java
@@ -1,3 +1,8 @@
+package duke;
+
+import duke.command.*;
+import duke.exception.DukeException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
