+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,0 +1,34 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
+import duke.task.Deadline;
+
+import java.time.LocalDateTime;
+
+public class DeadlineCommand extends Command {
+    String desc;
+    LocalDateTime datetime;
+
+    public DeadlineCommand(String desc, LocalDateTime datetime) {
+        this.desc = desc;
+        this.datetime = datetime;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        Deadline tmp = new Deadline(desc, datetime);
+        tasks.addDeadline(tmp);
+        storage.write(tasks.toStringWritable());
+        ui.showOutput("Got it. I added this task:");
+        ui.showOutput("\t" + tmp);
+        ui.showOutput("Now you have " + tasks.getLength() + " tasks in the list.");
+    }
+
+
+    public boolean isExit() {
+        return false;
+    }
+}
