+++ src/main/java/duke/command/EventCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
+import duke.task.Event;
+
 import java.time.LocalDateTime;
 
 public class EventCommand extends Command {
