+++ src/main/java/duke/command/GoodbyeCommand.java
@@ -1,0 +1,18 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
+
+public class GoodbyeCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+        ui.goodbye();
+    }
+
+    public boolean isExit() {
+        return true;
+    }
+}
