+++ src/main/java/duke/command/ListCommand.java
@@ -1,5 +1,12 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.exception.DukeException;
+
 public class ListCommand extends Command {
-    void execute(TaskList tasks, Ui ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         int len = tasks.getLength();
         ui.showOutput("You currently have " + len + " tasks:");
         ui.showOutput(tasks.toString());
