+++ src/main/java/duke/exception/DukeException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 /**
  * General Exception for the Duke bot.
  */
